requirements.txt
PyQt5>=5.15
openai==0.28.1
google.generativeai==0.8.5
cohere